fix: consolidate more validators into "common"
