+++ packages/validators/src/schemas/latest.ts
@@ -1,2 +1,13 @@
 export * from './v2023_09_28/experience';
 export { PatternFields } from './v2023_09_28/pattern';
+export { ComponentPropertyValueSchema } from '@/schemas/v2023_09_28/common';
+export { ComponentPropertyValue } from '@/schemas/v2023_09_28/common';
+export { PatternPropertySchema } from '@/schemas/v2023_09_28/common';
+export { PatternPropertiesSchema } from '@/schemas/v2023_09_28/common';
+export { BreakpointSchema } from '@/schemas/v2023_09_28/common';
+export { BaseComponentTreeNodeSchema } from '@/schemas/v2023_09_28/common';
+export { ComponentTreeNode } from '@/schemas/v2023_09_28/common';
+export { ComponentVariableSchema } from '@/schemas/v2023_09_28/common';
+export { ComponentTreeNodeSchema } from '@/schemas/v2023_09_28/common';
+export { ComponentTreeSchema } from '@/schemas/v2023_09_28/common';
+export { localeWrapper } from '@/schemas/v2023_09_28/common';
